requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==4.2.5
django-adminlte3==0.1.6
django-bootstrap-datepicker-plus==5.0.5
django-multiselectfield==0.1.12
django-unfold==0.20.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
fontawesomefree==6.5.1
mysql-connector-python==8.2.0
mysqlclient==2.2.0
Pillow==10.1.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
